+++ inc/GeneralThemeOptions/Controller.php
@@ -10,9 +10,10 @@ class Controller
 
   public const OPTION_NAME = 'sfx_general_options';
 
+
+
   public function __construct()
   {
-
     Settings::register(self::OPTION_NAME);
     AdminPage::register();
 
@@ -20,7 +21,6 @@ class Controller
     add_action('init', [$this,'handle_options']);
     add_action('update_option_' . self::OPTION_NAME, [$this, 'handle_options'], 10, 2);
   }
-
 
 
   public function handle_options(): void
@@ -118,6 +118,19 @@ class Controller
   {
     $options = get_option(self::OPTION_NAME, []);
     return !empty($options[$option_key]);
+  }
+
+  
+  public static function get_feature_config(): array
+  {
+    return [
+      'class' => self::class,
+      'menu_slug' => AdminPage::$menu_slug,
+      'page_title' => AdminPage::$page_title,
+      'description' => AdminPage::$description,
+      'error' => 'Missing GeneralThemeOptionsController class in theme',
+      'hook'  => null,
+    ];
   }
 
 }