+++ inc/wpoptimizer/Controller.php
@@ -674,4 +674,19 @@ class Controller
             return $open;
         }, 10, 2);
     }
+
+    public static function get_feature_config(): array
+    {
+        return [
+            'class' => self::class,
+            'menu_slug' => AdminPage::$menu_slug,
+            'page_title' => AdminPage::$page_title,
+            'description' => AdminPage::$description,
+            'activation_option_name' => 'sfx_general_options',
+            'activation_option_key' => 'enable_wp_optimizer',
+            'option_value' => true,
+            'hook' => null,
+            'error' => 'Missing WPOptimizerController class in theme',
+        ];
+    }
 }